requirements.txt
python=3.9
mlflow=1.20.2
numpy=1.20.3
scipy=1.6.2
matplotlib=3.4.2
gpytorch=1.5.1
